move the xml reading and decoding to a separate function
