+++ safe_rcm/product/utils.py
@@ -1,24 +1,8 @@
-import posixpath
-from urllib.parse import urlsplit, urlunsplit
-
 import toolz
 
 
 def is_scalar(x):
     return not toolz.itertoolz.isiterable(x) or isinstance(x, (str, bytes))
-
-
-def absolute_url_path(url):
-    """convert the url's path component to absolute"""
-    if url.count("::") > 1:
-        # TODO: unlike urllib.parse, `fsspec` allows nested urls
-        #       so we need to find a way to support that, as well
-        raise ValueError("don't know how to deal with nested urls")
-
-    split = urlsplit(url)
-    absolute_path = posixpath.abspath(split.path)
-
-    return urlunsplit(split._replace(path=absolute_path))
 
 
 def split_marked(mapping, marker="@"):
